Fixed a bug in reporting adherence in response to the back button.
Implemented synchronization on the tablet in response to documents being saved on the workstation.

Improved the BloodPressureReportingView so that the virtual keyboard does not appear on selecting a text input. Also, the NumberPad does not obscure any of the text inputs.

Added "Other Reason" to the NonadherenceReasonViews.

Temporarily fixed a bug in the VideoPlayer so that the playHead indicator does not get longer than the playHead. Currently it ends short of the end and needs to be fixed.