Updated the EquipmentScheduleItemTimelineView and MedicationScheduleItemTimelineView.
Implemented a MedicationColorSource that contains a dictionary for getting colors from ndc codes. It is a singleton implemented using the WorkstationKernel.

Updated the MedicationImageSkin and created a MedicationImage subclass so that a colored dot is shown in place of the image as it is loading and if the image fails to load.