Fixed a bug in AdherenceItemStitcher in that it did not have VitalSign as a required document.
Refactored the evaluation of adherence performance so that each plugin (Medications and Equipment) is responsible for making its own adherence performance evaluation by creating an AdherencePerformanceEvaluator through a new method in its ViewFactory. This new technique is working to produce simple string assertions, but the accuracy of these evaluations has not been tested and the visualization of the assertions with color and images has not been implemented.

Refactored the ScheduleModel so that getScheduleItemOccurrences is reusable and that duplicate code is eliminated.

Introduced constants in several cases where the number of milliseconds in a given time period was previously calculated when needed.