Updated the color of the goal region for the charts in the BloodPressureFullView. There are bugs in the drawing of the correct goal region on the correct chart.
Updated the AdherencePerformanceEvaluator classes to return an AdherencePerformanceAssertion instead of just a string. This assertion is used in the AdherencePerformanceAssertionItemRenderer to present an icon, string, and to assign red or green color to the string.

Fixed a bug in the AdherencePerformanceModel. It is necessary that the same ScheduleItemOccurrences are used for the clock and for the adherence performance. Otherwise local changes to adherence are not represented in the adherence performance.