Android.DeviceGateway now forces Wifi connection to stay on.
Fixed the glow behind check and x on the ScheduleClockWidgetView to be sized based on the size of the check or x.

Now ensures that blood pressures that are submitted will be saved to the record, regardless of whether the user continues through the sequence of views that appear when a blood pressure is in hypotension or severe hypertension range.

Updated the AdherencePerformanceWidgetView so that it can show single images when the images for each AdherencePerformanceAssertion are the same.