There were bugs in the handling of problems because they would be loaded for all of the records in clinician mode and then individually loaded and unloaded with each record. This resulted in duplicate problems and missing problems in the Select Record.
I opted for a quick fix for now in which problems are not loaded and unloaded with each record. They are just loaded once in clinician mode and never in patient mode. This is ok for now because problems are never used in the tablet app except in clinician mode. This results in one less server call.