Implemented the ability to not persist changes to the record including the addition of documents and relationships and the removal (voiding, deleting, archiving) of documents.
Created a subclass of ViewNavigator that fires an event every time that a view is popped, not just a single event for a series of pops. This is used to ensure that pops are synchronized between collaborating clients.

Refactored the SynchronizationService to have the flags _primaryCall and _initiatedLocally. Primary call is used to decide whether or not a call should be synchronized. Primary calls are synchronized while others are not. There is no longer the need for each method that uses synchronization to have a parameter, such as "calledLocally" that was used in the past. _initiatedLocally describes whether the call to a method, even if it passed through the collaboration server, was actually initiated locally. It is necessary to know this to decide whether data should be persisted or certain actions should be taken.