Made the hit target for the check box for a schedule health action larger and made the hit target for the entire health action item non-intersecting with this check box hit target. That way, when the check box is disabled, clicking on the check box does not select the item.
Added a save button to the BloodGlucoseHealthActionInputView because the save button in the action bar is too small for some users.

Added a comment in each blood glucose VitalSign to indicate if the result was self-reported or from a device. This is visualized in the BloodGlucoseHistoryView with each BloodGlucoseItemRenderer showing the wireless icon for VitalSigns that were from a device.